Draft new NodeClient handle for ouroborus
